+++ tests/service/e2e/cucumber/steps/envelope_steps.go
@@ -4,12 +4,15 @@ import (
 	"context"
 	"fmt"
 	"math/rand"
+	"net/http"
 	"reflect"
 	"regexp"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
+
+	"github.com/cenkalti/backoff/v4"
 
 	"github.com/Shopify/sarama"
 	"github.com/cucumber/godog"
@@ -333,6 +336,29 @@ func (sc *ScenarioContext) iRegisterTheFollowingChains(table *gherkin.PickleStep
 
 		// set aliases
 		sc.aliases.Set(res, sc.Pickle.Id, utilsCols.Rows[i+1].Cells[0].Value)
+
+		chainRegURL, _ := sc.aliases.Get(alias.GlobalAka, "chain-registry")
+		req, err := http.NewRequest("GET", fmt.Sprintf("%s/%s", chainRegURL, res.UUID), nil)
+		if err != nil {
+			return err
+		}
+
+		err = backoff.RetryNotify(
+			func() error {
+				_, err2 := sc.httpClient.Do(req)
+				return err2
+			},
+			backoff.WithMaxRetries(backoff.NewConstantBackOff(time.Second), 5),
+			func(err error, duration time.Duration) {
+				log.WithFields(log.Fields{
+					"chain_uuid": res.UUID,
+				}).WithError(err).Debug("scenario: chain proxy is still not ready")
+			},
+		)
+
+		if err != nil {
+			return err
+		}
 	}
 
 	return nil